# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CINTEXA NEXUS AI Business Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List support tickets
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: priority
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    patch:
      operationId: updateTicket
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    delete:
      operationId: deleteTicket
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Marketing Campaigns ────────────────────────────────────────────────
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Tasks ──────────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - in: query
          name: projectId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── AI Assistant ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Ticket:
      type: object
      required: [id, subject, status, priority, channel, contactName, createdAt]
      properties:
        id:
          type: integer
        subject:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, waiting, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        channel:
          type: string
          enum: [email, chat, phone, social, portal]
        contactName:
          type: string
        assignedTo:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        firstRespondedAt:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        sla:
          $ref: "#/components/schemas/SlaStatus"
    SlaStatus:
      type: object
      required: [firstResponseDueAt, resolutionDueAt, firstResponseBand, resolutionBand, minutesToResolutionBreach]
      properties:
        firstResponseDueAt: { type: string }
        resolutionDueAt: { type: string }
        firstResponseBand:
          type: string
          enum: [met, on_track, at_risk, breached]
        resolutionBand:
          type: string
          enum: [met, on_track, at_risk, breached]
        minutesToResolutionBreach:
          type: ["number", "null"]
...
    TicketInput:
      type: object
      required: [subject, priority, channel, contactName]
      properties:
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, waiting, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        channel:
          type: string
          enum: [email, chat, phone, social, portal]
        contactName:
          type: string
        assignedTo:
          type: string
    TicketUpdate:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, waiting, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedTo:
          type: string

    # ── Campaign ─────────────────────────────────────────────────────────
...
    Project:
      type: object
      required: [id, name, status, progress, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        progress:
          type: integer
        owner:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        taskCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        owner:
          type: string
        dueDate:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        progress:
          type: integer
        owner:
          type: string
        dueDate:
          type: string

    # ── Task ─────────────────────────────────────────────────────────────
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: ["integer", "null"]
        assignedTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: integer
        assignedTo:
          type: string
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedTo:
          type: string
        dueDate:
          type: string

    # ── AI ───────────────────────────────────────────────────────────────
...